Startup script updates
The script itself is functional, but the parameters don't actually function.
So, this startup script will actually start an environment, but all of the
configuration is done within the file (rather than on the CLI with exporting
values).

So, still a step in the right direction, but more work remains.